feat: refactor form component #84
